requirements.txt
requests
beautifulsoup4
markdownify
aiohttp
pdfplumber
python-docx
pytesseract
Pillow
psutil
